Soil moisture sensor
It is a measure of gravimetric soil moisture of the soil. Soil moisture plays vital role in plant growth and provides them nutrients. Therefore, it is important to detect the soil moisture content for proper development of plants. This IoT project uses soil moisture sensor to notify the user when the soil gets too dry or too wet. This device is developed on Arduino platform and has the capability to communicate in a wireless manner with the other devices to provide the soil moisture value.